Consider N walls, each of unit width, situated next to each other.
You have to select any two walls i and j (1<=i, j<=N) such that if you break all walls except i and j and fill the spaces between them with water, then the amount of water stored is maximum.

Write a program to find the maximum amount of water that can be accumulated between the walls.

eg: 1 2 3 4 5

Ans: we choose wall 2 and 5, the amount of water that can be stored will be the distance between walls i.e (5-2-1) = 4 which is maximum